# src/backend/openapi.yaml
openapi: 3.1.0
info:
  title: LiDAR Forest Analysis API
  version: 1.0.0
  description: |
    REST API for the LiDAR Forest Analysis Platform. Transform LiDAR point cloud data
    into professional forest inventory reports, carbon stock estimates, and 3D visualizations.

    ## Authentication

    All API requests require authentication via API key. Include your API key in requests using
    one of these methods:

    - **Authorization header** (recommended): `Authorization: Bearer lf_live_xxx`
    - **X-API-Key header**: `X-API-Key: lf_live_xxx`

    ## Rate Limits

    Rate limits vary by tier:

    | Tier | Requests/Minute | Requests/Day |
    |------|-----------------|--------------|
    | Free | 60 | 10,000 |
    | Starter | 120 | 50,000 |
    | Professional | 300 | 200,000 |
    | Enterprise | 1,000 | Unlimited |

    Rate limit headers are included in all responses:
    - `X-RateLimit-Limit`: Max requests per minute
    - `X-RateLimit-Remaining`: Remaining requests this minute
    - `X-RateLimit-Reset`: Time when the rate limit resets (ISO 8601)
  contact:
    name: API Support
    email: api@lidarforest.com
    url: https://docs.lidarforest.com
  license:
    name: Proprietary
    url: https://lidarforest.com/terms
servers:
  - url: https://api.lidarforest.com/api/v1
    description: Production server
  - url: https://staging-api.lidarforest.com/api/v1
    description: Staging server
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /analyses:
    get:
      summary: List analyses
      tags: [Analyses]
      operationId: listAnalyses
      parameters:
        - name: projectId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, PROCESSING, COMPLETED, FAILED]
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/AnalysisType'
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of analyses
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisListResponse'
    post:
      summary: Start analysis
      tags: [Analyses]
      operationId: createAnalysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAnalysisInput'
      responses:
        '201':
          description: Analysis started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /analyses/{analysisId}:
    get:
      summary: Get analysis
      tags: [Analyses]
      operationId: getAnalysis
      parameters:
        - $ref: '#/components/parameters/AnalysisId'
      responses:
        '200':
          description: Analysis details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Cancel/delete analysis
      tags: [Analyses]
      operationId: deleteAnalysis
      parameters:
        - $ref: '#/components/parameters/AnalysisId'
      responses:
        '200':
          description: Analysis cancelled/deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /analyses/{analysisId}/results:
    get:
      summary: Get analysis results
      tags: [Analyses]
      operationId: getAnalysisResults
      parameters:
        - $ref: '#/components/parameters/AnalysisId'
      responses:
        '200':
          description: Analysis results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResultsResponse'
        '400':
          description: Analysis not completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /analyses/{analysisId}/trees:
    get:
      summary: Get detected trees
      tags: [Inventory]
      operationId: getAnalysisTrees
      parameters:
        - $ref: '#/components/parameters/AnalysisId'
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of detected trees
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreeListResponse'
  /analyses/{analysisId}/stands:
    get:
      summary: Get stand summaries
      tags: [Inventory]
      operationId: getAnalysisStands
      parameters:
        - $ref: '#/components/parameters/AnalysisId'
      responses:
        '200':
          description: Stand summaries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandListResponse'

  # Reports
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key as Bearer token
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key header
  parameters:
...
    AnalysisId:
      name: analysisId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    AnalysisType:
      type: string
      enum: [TREE_DETECTION, SPECIES_CLASSIFICATION, CARBON_ESTIMATE, FULL_INVENTORY]
...
    Status:
      type: string
      enum: [PENDING, PROCESSING, COMPLETED, FAILED]

    # Request schemas
...
    CreateAnalysisInput:
      type: object
      required: [projectId, name, type, fileIds]
      properties:
        projectId:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 200
        type:
          $ref: '#/components/schemas/AnalysisType'
        fileIds:
          type: array
          items:
            type: string
            format: uuid
          minItems: 1
        parameters:
          type: object
          description: Analysis-specific parameters
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        hasMore:
          type: boolean
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
...
    AnalysisSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          $ref: '#/components/schemas/AnalysisType'
        status:
          $ref: '#/components/schemas/Status'
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
    Analysis:
      allOf:
        - $ref: '#/components/schemas/AnalysisSummary'
        - type: object
          properties:
            projectId:
              type: string
              format: uuid
            progress:
              type: integer
              minimum: 0
              maximum: 100
            errorMessage:
              type: string
    AnalysisListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Analysis'
        pagination:
          $ref: '#/components/schemas/Pagination'
    AnalysisResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Analysis'
        message:
          type: string
    AnalysisDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          allOf:
            - $ref: '#/components/schemas/Analysis'
            - type: object
              properties:
                parameters:
                  type: object
                startedAt:
                  type: string
                  format: date-time
                files:
                  type: array
                  items:
                    type: object
                    properties:
                      id:
                        type: string
                      filename:
                        type: string
    AnalysisResultsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            analysisId:
              type: string
              format: uuid
            type:
              $ref: '#/components/schemas/AnalysisType'
            results:
              type: object
    Tree:
      type: object
      properties:
        id:
          type: string
          format: uuid
        species:
          type: string
        dbh:
          type: number
          description: Diameter at breast height (cm)
        height:
          type: number
          description: Tree height (m)
        crownDiameter:
          type: number
          description: Crown diameter (m)
        latitude:
          type: number
        longitude:
          type: number
        confidence:
          type: number
          minimum: 0
          maximum: 1
        carbonStock:
          type: number
          description: Carbon stock (kg)
    TreeListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Tree'
        pagination:
          $ref: '#/components/schemas/Pagination'
    Stand:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        areaHectares:
          type: number
        treeCount:
          type: integer
        basalArea:
          type: number
        volumePerHectare:
          type: number
        dominantSpecies:
          type: string
        meanDbh:
          type: number
        meanHeight:
          type: number
        carbonStockPerHectare:
          type: number
    StandListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Stand'
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...